Factored out assembly_utils.rs and definitions/constants.rs
