+++ name-as/src/assembler/mod.rs
@@ -2,6 +2,5 @@ pub mod assemble_file;
 pub(crate) mod assemble_instruction;
 pub mod assembler;
 pub(crate) mod assembly_helpers;
-pub(crate) mod assembly_utils;
 pub(crate) mod directive_handler;
 pub(crate) mod directives;
