+++ name-as/src/definitions/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod constants;
 pub(crate) mod expandables;
 pub(crate) mod implementations;
 pub(crate) mod pseudo_instructions;
